+++ features/fixtures/laravel58/app/Jobs/HandledJob.php
@@ -2,14 +2,13 @@
 
 namespace App\Jobs;
 
-use Bugsnag\BugsnagLaravel\Facades\Bugsnag;
 use Exception;
 use Illuminate\Bus\Queueable;
-use Illuminate\Contracts\Queue\ShouldBeUnique;
+use Illuminate\Queue\SerializesModels;
+use Illuminate\Queue\InteractsWithQueue;
 use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Foundation\Bus\Dispatchable;
-use Illuminate\Queue\InteractsWithQueue;
-use Illuminate\Queue\SerializesModels;
+use Bugsnag\BugsnagLaravel\Facades\Bugsnag;
 
 class HandledJob implements ShouldQueue
 {