+++ features/fixtures/laravel66/app/Jobs/UnhandledJob.php
@@ -3,15 +3,14 @@
 namespace App\Jobs;
 
 use Bugsnag\BugsnagLaravel\Facades\Bugsnag;
-use Exception;
 use Illuminate\Bus\Queueable;
-use Illuminate\Contracts\Queue\ShouldBeUnique;
 use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Foundation\Bus\Dispatchable;
 use Illuminate\Queue\InteractsWithQueue;
 use Illuminate\Queue\SerializesModels;
+use RuntimeException;
 
-class HandledJob implements ShouldQueue
+class UnhandledJob implements ShouldQueue
 {
     use Dispatchable, InteractsWithQueue, Queueable, SerializesModels;
 
@@ -24,6 +23,6 @@ class HandledJob implements ShouldQueue
     {
         Bugsnag::leaveBreadcrumb(__METHOD__);
 
-        Bugsnag::notifyException(new Exception('Handled :)'));
+        throw new RuntimeException('uh oh :o');
     }
 }