+++ features/lib/laravel.rb
@@ -1,5 +1,7 @@
 require 'net/http'
 require 'yaml'
+require 'json'
+require 'tempfile'
 
 class Laravel
   class << self
@@ -27,8 +29,13 @@ class Laravel
     end
 
     def major_version
-      # e.g. laravel56 -> 5, lumen8 -> 8
-      Integer(/^(?:laravel|lumen)(\d)/.match(fixture)[1])
+      # the first "canonical segment" is the first digit of the version number,
+      # aka the major version
+      version.canonical_segments.first
+    end
+
+    def version
+      @version ||= load_version_from_fixture
     end
 
     def lumen?
@@ -46,6 +53,27 @@ class Laravel
       true
     end
 
+    # the command to run the queue worker for a single job
+    def queue_worker_once_command(tries)
+      if version < "5.3.0"
+        "php artisan queue:work --tries=#{tries}"
+      else
+        "php artisan queue:work --once --tries=#{tries}"
+      end
+    end
+
+    # the command to run the queue worker as a daemon
+    def queue_worker_daemon_command(tries)
+      # the command to run the queue worker was 'queue:listen' but changed to
+      # 'queue:work' in Laravel 5.3 ('queue:work' exists on older Laravels, but
+      # is not quite equivalent)
+      if version < "5.3.0"
+        "php artisan queue:listen --tries=#{tries}"
+      else
+        "php artisan queue:work --tries=#{tries}"
+      end
+    end
+
     private
 
     def load_port_from_docker_compose
@@ -53,6 +81,32 @@ class Laravel
       service = compose_file.fetch("services").fetch(ENV['LARAVEL_FIXTURE'])
 
       service.fetch("ports").first.fetch("published")
+    end
+
+    def load_version_from_fixture
+      # get and parse the composer.lock file from the fixture
+      composer_lock = Tempfile.create("#{fixture}-composer.lock") do |file|
+        # copy the composer lock file out of the fixture so we can read it
+        Maze::Docker.cp(fixture, source: "/app/composer.lock", destination: file.path)
+
+        # 'file.read' won't reflect the changes made by docker cp, so we use
+        # JSON.load_file to reload the file & parse it
+        JSON.load_file(file.path)
+      end
+
+      framework_section = composer_lock["packages"].find { |package| package["name"] == framework_package_name }
+      version = framework_section["version"].delete_prefix("v")
+
+      Gem::Version.new(version)
+    end
+
+    # the composer package name of the framework being used (Lumen or Laravel)
+    def framework_package_name
+      if lumen?
+        "laravel/lumen-framework"
+      else
+        "laravel/framework"
+      end
     end
   end
 end