+++ features/steps/laravel_steps.rb
@@ -14,22 +14,41 @@ When(/^I start the laravel fixture$/) do
   }
 end
 
-# TODO: contribute this back to Maze Runner
-#       https://github.com/bugsnag/maze-runner/pull/425
 module Maze
   class Docker
     class << self
+      # TODO: remove when https://github.com/bugsnag/maze-runner/pull/425 is merged
       def exec(service, command, detach: false)
         flags = detach ? "--detach" : ""
 
         run_docker_compose_command("exec #{flags} #{service} #{command}")
+      end
+
+      # TODO: contribute this back to Maze Runner
+      #       probably need a nicer API, capable of doing a copy in either
+      #       direction (right now this can only copy from the service to the
+      #       local machine)
+      def cp(service, source:, destination:)
+        run_docker_compose_command("cp #{service}:#{source} #{destination}")
       end
     end
   end
 end
 
 When("I start the laravel queue worker") do
-  Maze::Docker.exec(Laravel.fixture, "php artisan queue:work", detach: true)
+  step("I start the laravel queue worker with --tries=1")
+end
+
+When("I start the laravel queue worker with --tries={int}") do |tries|
+  Maze::Docker.exec(Laravel.fixture, Laravel.queue_worker_daemon_command(tries), detach: true)
+end
+
+When("I run the laravel queue worker") do
+  step("I run the laravel queue worker with --tries=1")
+end
+
+When("I run the laravel queue worker with --tries={int}") do |tries|
+  Maze::Docker.exec(Laravel.fixture, Laravel.queue_worker_once_command(tries))
 end
 
 When("I navigate to the route {string}") do |route|
@@ -98,4 +117,68 @@ Then("the session payload field {string} matches the current major Lumen version
   end
 
   step("the session payload field '#{path}' matches the regex '^((\\d+\\.){2}\\d+|\\d\\.x-dev)$'")
+end
+
+# TODO: remove when https://github.com/bugsnag/maze-runner/pull/433 is released
+Then("the event has {int} breadcrumb(s)") do |expected|
+  breadcrumbs = Maze::Server.errors.current[:body]['events'].first['breadcrumbs']
+
+  Maze.check.equal(
+    expected,
+    breadcrumbs.length,
+    "Expected event to have '#{expected}' breadcrumbs, but got: #{breadcrumbs}"
+  )
+end
+
+Then("the event has no breadcrumbs") do
+  breadcrumbs = Maze::Server.errors.current[:body]['events'].first['breadcrumbs']
+
+  Maze.check.true(
+    breadcrumbs.nil? || breadcrumbs.empty?,
+    "Expected event not to have breadcrumbs, but got: #{breadcrumbs}"
+  )
+end
+
+# conditionally run a step if the laravel version matches a specified version
+#
+# e.g. this will only check app.type on Laravel 5.2 and above:
+#      on Laravel versions > 5.1 the event "app.type" equals "Queue"
+Then(/^on Laravel versions (>=?|<=?|==) ([0-9.]+) (.*)$/) do |operator, version, step_to_run|
+  should_run_step = Laravel.version.send(operator, version)
+
+  # make sure this step is debuggable!
+  $logger.debug("Laravel v#{Laravel.version} #{operator} #{version}? #{should_run_step}")
+
+  if should_run_step
+    step(step_to_run)
+  else
+    $logger.info("Skipping step on Laravel v#{Laravel.version}: #{step_to_run}")
+  end
+end
+
+# conditionally run a number of steps if the laravel version matches a specified version
+#
+# e.g. this will only run the indented steps on Laravel 5.2 and above:
+#      on Laravel versions > 5.1:
+#         """
+#         the event "app.type" equals "Queue"
+#         the event "other.thing" equals "yes"
+#         """
+Then(/^on Laravel versions (>=?|<=?|==) ([0-9.]+):/) do |operator, version, steps_to_run|
+  should_run_steps = Laravel.version.send(operator, version)
+
+  # make sure this step is debuggable!
+  $logger.debug("Laravel v#{Laravel.version} #{operator} #{version}? #{should_run_steps}")
+
+  if should_run_steps
+    steps_to_run.each_line(chomp: true) do |step_to_run|
+      step(step_to_run)
+    end
+  else
+    indent = " " * 4
+    # e.g. "a step\nanother step\n" -> "    1) a step\n    2) another step"
+    steps_indented = steps_to_run.each_line.map.with_index(1) { |step, i| "#{indent}#{i}) #{step.chomp}" }.join("\n")
+
+    $logger.info("Skipping steps on Laravel v#{Laravel.version}:\n#{steps_indented}")
+  end
 end